move stuff out of function
